requirements.txt
# Web 框架
Flask>=2.0

# OpenAI API
openai>=1.0.0

# 音頻處理 (即時錄音功能)
# 注意：在 macOS 和 Linux 上可能需要額外的系統級依賴
pyaudio>=0.2.11

# 環境變數管理
python-dotenv>=0.19.0


# ===============================
#  開發工具 (Development Tools)
# ===============================

# 程式碼風格檢查器
flake8>=5.0.0

# 程式碼自動格式化工具
black>=22.0.0